+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [1.1.0] - 2024-08-27
+
+- TinyMCE Editor for Jmix 2.3.2
+- #1 Upgrade to latest vaadin addon version
+- #2 Make TinyMceEditor configure method return TinyMceEditor instance and not TinyMCE
+
 ## [1.0.0] - 2024-05-01
 
-- DurationField for Jmix 2.2.1
+- TinyMCE Editor for Jmix 2.2.1